turn search field into separate component
